Implemented a Tag service container
Much of the tag controller's functionality has now been offloaded to a dedicated service container. This allows the methods to be much easier to write and allows them to be used from any controller, as opposed to just the TagsController.

The GetAllChildren method has also been rewritten with a recursive helper function and a hashset. This hasn't been tested yet but should hopefully be functional out of the box.