Fixed a minor issue of no content found being unhandled when getting child tags
When getting all the children of a tag to be displayed in the tag menu, the check that the status is ok was somehow getting bypassed. This meant that the program was trying to get the json from an empty response, causing an error in the console. It had no adverse side-effects, but it's now been fixed regardless.