Fixed adding and deleting tags from the graph
It appears that the solution I found using d3Force('link) to separate the nodes breaks the force graph? I'm guessing it converts it to a d3 object instead of a force graph one or something. Either way I'm planning on going full d3 eventually  so the nodes will just have to stay clumped together for the time being to ensure that adding and removing tags still works.