Adding and removing tag relations now updates the associated visual elements
When adding or removing a tag relation, it will now reload all parents and children. This ensures that any new or removed tags are present with all the others. It's probably slower than just appending the individual tag visual, but it's quick and easy to implement and relies on the backend rather than the frontend which avoids parity issues between the two.